Removed document eventlistener for keydown and directly listens on the gridRootEelement Renamed useKeyboardControls.tsx to useKeyboard Organized keyboard folder with utils folder
