+++ src/keyboard/index.ts
@@ -1,3 +1,4 @@
 export * from './types'
-export * from './navigation.utils'
-export * from './useKeyboardControls'
+export * from './utils'
+export * from './useKeyboard'
+export * from './constants'
