--- src/keyboard/reservedKeyCodes.ts
@@ -1,1 +1,0 @@
-export const RESERVED_KEY_CODES = ['ArrowLeft', 'ArrowRight', 'ArrowDown', 'ArrowUp', 'Tab']
