+++ src/keyboard/utils/getDefaultValueFromValue.util.ts
